+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## 1.5.3
+## Fixed
+- `genPdfCover()` linux platfrom -> override gen png
+- `windowManagerensureInitialized()` methodUnimplementedError() -> fixed
+- `isStoragePermissionGranted()` methodUnimplementedError() -> fixed
+
 ## 1.5.1
 ### Added
 - added for linux,
-- ensure do it. `await ThanPkg.windowManagerensureInitialized();`
-- `await ThanPkg.platform.toggleFullScreen(isFullScreen: true);`
+- ensure do it. `windowManagerensureInitialized();`
+- `toggleFullScreen(isFullScreen: true);`